python-ml-service/requirements.txt
fastapi
uvicorn
pydantic
pandas
prophet
requests
python-dotenv
openai
gunicorn